pyproject.toml
[project]
name = "comfyui-ap-clipseg-light"
description = "Lightweight ComfyUI node for text-prompted segmentation masks using CLIPSeg. numpy 2.x safe, TensorFlow-free, auto model caching."
version = "1.1.0"
license = { file = "LICENSE" }
readme = "README.md"
dependencies = ["torchvision>=0.15", "transformers>=4.30", "Pillow>=9.0"]

[project.urls]
Repository = "https://github.com/adampolczynski/ComfyUI_AP_ClipSEG_Light"

[tool.comfy]
PublisherId = "adampolczynski"
DisplayName = "AP ClipSEG Light"
Icon = "icon.ico"
